+++ CHANGELOG.md
@@ -1,7 +1,10 @@
-# 0.21.9
+# Current
 
 - Add utility function `get_onchain_price()` to ask on-chain price of a 
   Uniswap v3 pool at any given block number
+- Add MEV blocking support in the form of `eth_defi.mev_blocker.MEVBlockerProvider`
+- Add JSON-RPC fallback switching in the form of `eth_defi.fallback_provider.FallbackProvider`
+- Add `HotWallet.create_for_testing`
 
 # 0.21.8